小说翻译v1.0/requirements.txt
# requirements.txt
PyQt6>=6.4.3
openai>=1.3.1
python-docx>=0.6.3
chardet>=4.0.0
logging>=0.5.1.2
hashlib>=4.3.2
uuid>=1.30
Pygments>=2.16.1